fix: remove fatal logs
