+++ clientset/get_clientset.go
@@ -1,6 +1,7 @@
 package clientset
 
 import (
+	"log"
 	kubernetesconfig "wheelie/kubernetes_config"
 
 	"k8s.io/client-go/kubernetes"
@@ -11,7 +12,7 @@ var Context string
 func GetClientSet() *kubernetes.Clientset {
 	clientset, err := kubernetes.NewForConfig(kubernetesconfig.GetConfig(Context))
 	if err != nil {
-		panic(err.Error())
+		log.Print(err.Error())
 	}
 
 	return clientset
